feat(Frontend): Updated authentication and improved search results display ✨
Implemented URL routing for login and signup pages using Flask's `url_for` function.  This ensures proper routing and prevents hardcoded URLs. 🚀

Added hover effect to product images in the search results.  This enhances user experience by showcasing additional product details on hover.  Now, the second image is displayed as a hover effect on the search results, or, for products with only one image, the same image is used.  This allows for more appealing and interactive presentation. 🖼️

This also fixes a subtle display issue with the search results and now the product rating is displayed correctly. 💯

This refactor improves the overall user experience, making the site more organized, user-friendly and professional. 🎉